Fixing an issue with calculations involving eccentricity when the semi major axis was something other than 1AU.
(i/periodFractions) is the proportion of the period that the planet has gone through. This needs to be multiplied by the period to find the actual time. The previous method gave the correct results since the period was 1**(3/2) which is 1 however for semi major axes other than 1, the results ho wrong.